+++ test/Main.hs
@@ -25,6 +25,7 @@ import qualified Unit.BuildEnvCompose
 import qualified Unit.BuildPlanOrder
 import qualified Unit.CliParser
 import qualified Unit.Components
+import qualified Unit.CppMacros
 import qualified Unit.Deps
 import qualified Unit.Doctor
 import qualified Unit.EmbeddedAssets
@@ -94,6 +95,7 @@ allTests = testGroup "hypha"
   , Unit.BuildPlanOrder.tests
   , Unit.CliParser.tests
   , Unit.Components.tests
+  , Unit.CppMacros.tests
   , Unit.Deps.tests
   , Unit.Doctor.tests
   , Unit.EmbeddedAssets.tests